footer div, routes, fix product prices
added footer div
added a route for the shopping cart
added basic oderitem and order models (commented out)
fixed routes for product details page
fixed the product prices decimal point

+++ django_ecommerce_site/webshop/models.py
@@ -7,7 +7,7 @@ class Product(models.Model):
     title = models.CharField(max_length=64)
     description = models.TextField()
     price = models.DecimalField(max_digits=6, decimal_places=2)
-    old_price = models.FloatField()
+    old_price = models.DecimalField(max_digits=6, decimal_places=2)
     image = models.CharField(max_length=256)
     stock = models.IntegerField(null=True)
     date_added = models.DateTimeField(default=timezone.now)
@@ -28,3 +28,14 @@ class User(models.Model):
 
     phone_number = models.CharField(max_length=16)
     date_of_birth = models.DateField()
+
+# class OrderItem():
+#     product = models.OneToOneField(Product, on_delete=models.SET_NULL)
+#     date_added = models.DateTimeField(auto_now=True)
+
+#     def __str__(self):
+#         return self.product.name
+
+# class Order(models.Model):
+#     ref_code = models.CharField(max_length=16)
+#     items = models.ManyToManyField(OrderItem)
